tests/submission/Google/video/controls/controls_property_exists.html

<!doctype html>
<html>
 <head>
  <title>video.controls - existence test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-controls">spec reference</a></p>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  "controls" in document.getElementById("video"),
  "videoElement.controls property exists");
});
  </script>
 </body>
</html>
